refactor : MVC 패턴으로 패키지 리펙토링
