+++ src/main/java/lotto/main/LottoApplication.java
@@ -1,5 +1,9 @@
-package lotto;
+package lotto.main;
 
+import lotto.controller.MatchResult;
+import lotto.domain.Buyer;
+import lotto.domain.Lotto;
+import lotto.domain.LottoTicket;
 import lotto.view.InputView;
 import lotto.view.OutputView;
