+++ src/main/java/lotto/view/OutputView.java
@@ -1,8 +1,7 @@
 package lotto.view;
 
-import lotto.LottoTicket;
+import lotto.domain.LottoTicket;
 
-import java.util.ArrayList;
 import java.util.List;
 
 public class OutputView {
